+++ app/routes/components/Button/changelog.codegen.ts
@@ -10,4 +10,6 @@ export const changelog = ''
   + '- 2025-09-18: Support the prop `size` (`sm`, `md` and `lg`).\n'
   + "- 2025-09-23: Update the default button's border color. Add `justify-content:\n"
   + '  center` to the button styles.\n'
+  + '- 2025-09-29: Make the prop `size` support `xs` and `xl` values & use tokens\n'
+  + '  from ThemeProvider.\n'
   ;