+++ src/components/Button/CHANGELOG.md
@@ -6,3 +6,5 @@
 - 2025-09-18: Support the prop `size` (`sm`, `md` and `lg`).
 - 2025-09-23: Update the default button's border color. Add `justify-content:
   center` to the button styles.
+- 2025-09-29: Make the prop `size` support `xs` and `xl` values & use tokens
+  from ThemeProvider.